rearrange mock notify declaration
