+++ backend/app/notify/notify_mock.go
@@ -2,14 +2,11 @@ package notify
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sync"
 	"time"
 
 	log "github.com/go-pkgz/lgr"
-
-	"github.com/umputun/remark/backend/app/store"
 )
 
 type MockDest struct {
@@ -41,23 +38,3 @@ func (m *MockDest) Get() []Request {
 	return res
 }
 func (m *MockDest) String() string { return fmt.Sprintf("mock id=%d, closed=%v", m.id, m.closed) }
-
-type mockStore struct {
-	data  map[string]store.Comment
-	email bool
-}
-
-func (m mockStore) Get(_ store.Locator, id string, _ store.User) (store.Comment, error) {
-	res, ok := m.data[id]
-	if !ok {
-		return store.Comment{}, errors.New("no such id")
-	}
-	return res, nil
-}
-
-func (m mockStore) GetUserEmail(_ store.Locator, userID string) (string, error) {
-	if !m.email {
-		return "", errors.New("no such user")
-	}
-	return "test@example.org", nil
-}
